+++ src/components/calcite-flow-item/calcite-flow-item.tsx
@@ -1,4 +1,4 @@
-import { Component, Element, Event, EventEmitter, Host, Prop, h } from "@stencil/core";
+import { Component, Element, Event, EventEmitter, Host, Listen, Prop, h } from "@stencil/core";
 import { VNode } from "@stencil/core/internal";
 import { focusElement, getElementDir } from "../utils/dom";
 import classnames from "classnames";
@@ -11,6 +11,7 @@ const SUPPORTED_ARROW_KEYS = ["ArrowUp", "ArrowDown"];
 
 /**
  * @slot menu-actions - A slot for adding `calcite-action`s to a menu under the `...` in the header. These actions are displayed when the menu is open.
+ * @slot fab - A slot for adding a `calcite-fab` (floating action button) to perform an action.
  * @slot footer-actions - A slot for adding `calcite-button`s to the footer.
  * @slot - A slot for adding content to the flow item.
  */
@@ -116,6 +117,12 @@ export class CalciteFlowItem {
 
   @Event() calciteFlowItemBackClick: EventEmitter;
 
+  /**
+   * Emitted when the content has been scrolled.
+   */
+
+  @Event() calciteFlowItemScroll: EventEmitter;
+
   // --------------------------------------------------------------------------
   //
   //  Private Properties
@@ -129,6 +136,12 @@ export class CalciteFlowItem {
   //  Private Methods
   //
   // --------------------------------------------------------------------------
+
+  @Listen("calcitePanelScroll")
+  handleCalcitePanelScroll(event: CustomEvent): void {
+    event.stopPropagation();
+    this.calciteFlowItemScroll.emit();
+  }
 
   queryActions(): HTMLCalciteActionElement[] {
     return Array.from(this.el.querySelectorAll(`[slot=${SLOTS.menuActions}] calcite-action`));
@@ -304,8 +317,9 @@ export class CalciteFlowItem {
   renderHeaderActions(): VNode {
     const menuActionsNode = this.el.querySelector(`[slot=${SLOTS.menuActions}]`);
 
-    const hasMenuActionsInBlacklisted =
-      menuActionsNode && menuActionsNode.closest(BLACKLISTED_MENU_ACTIONS_COMPONENTS.join(","));
+    const hasMenuActionsInBlacklisted = menuActionsNode?.closest(
+      BLACKLISTED_MENU_ACTIONS_COMPONENTS.join(",")
+    );
 
     const hasMenuActions = !!menuActionsNode && !hasMenuActionsInBlacklisted;
     const actionCount = hasMenuActions ? menuActionsNode.childElementCount : 0;
@@ -352,6 +366,15 @@ export class CalciteFlowItem {
     ) : null;
   }
 
+  renderFab(): VNode {
+    const hasFab = this.el.querySelector(`[slot=${SLOTS.fab}]`);
+    return hasFab ? (
+      <div class={CSS.fabContainer} slot={PANEL_SLOTS.fab}>
+        <slot name={SLOTS.fab} />
+      </div>
+    ) : null;
+  }
+
   render() {
     const { el } = this;
     const dir = getElementDir(el);
@@ -369,6 +392,7 @@ export class CalciteFlowItem {
           {this.renderHeaderActions()}
           <slot />
           {this.renderFooterActions()}
+          {this.renderFab()}
         </calcite-panel>
       </Host>
     );