+++ src/components/calcite-flow-item/resources.ts
@@ -11,11 +11,13 @@ export const CSS = {
   menuContainer: "menu-container",
   menuButton: "menu-button",
   menu: "menu",
-  menuOpen: "menu--open"
+  menuOpen: "menu--open",
+  fabContainer: "fab-container"
 };
 
 export const SLOTS = {
   menuActions: "menu-actions",
+  fab: "fab",
   footerActions: "footer-actions"
 };