+++ src/components/interfaces.ts
@@ -10,6 +10,6 @@ export type CalciteBlockSectionToggleDisplay = "button" | "switch";
 
 export type FlowDirection = "advancing" | "retreating";
 
-export type CalciteAppearance = "solid" | "clear";
+export type CalciteAppearance = "solid" | "clear" | "outline";
 
 export type CalciteScale = "s" | "m" | "l";